Add Publish plugin
